fix: float64 window reduction, provenance in the T-D test, honest gate list
Two failures my own changes caused, and one stale gate.

The window porosity reduction was in float32. A window is 4096 cells, so a
float64 reduction costs nothing and the float32 one differed from the exact
value by 7e-6 relative -- enough to move cond_por in the sixth decimal and
to fail an analytically-exact test for a reason that is not the model.

The T-D pointer test pinned the three leaked correlation lengths, so it
would have passed just as happily on the superseded all-80-volume fit --
which is precisely the leak it now exists to prevent. It asserts the
PROVENANCE first: split == train, the split_v3 index, 58 volumes, 1 598 000
patches. A second test pins which axes have a usable length (z and x yes, y
no) so the y artefact cannot quietly become a fact.

DEVELOPMENT.md claimed exactly five tests fail and everything else passes. A
full run gives twelve. Six more live in test_downstream_utility.py and are
NOT new: they have failed since the augmentation arms were added in 63cfe0b
to a test file last written before them in f9e1afa, and the list was never
updated. Five trip because Arm.total_patches is the absolute constant 32 000,
which silently assumes budget.patch_count == 16000 -- so a toy budget cannot
express an augmentation arm at all; expressing the override as a MULTIPLE of
patch_count would fix all five and remove the coupling. The sixth asserts Arm
has two fields when it has had five for some time. Recorded rather than
fixed, because they are outside this review's scope and the gate has to be
true before it is useful.

+++ src/poregen/diffusion/sampler.py
@@ -1425,7 +1425,13 @@ class VolumeGenerator:
                 phi = por_default * float(block.mean())
             else:
                 cells = window_tile_cells(ov, P, L, local_por_map, por_default)
-                phi = float((cells * block).mean())
+                # float64 for the reduction: a window is only L**3 = 4096
+                # cells, so the cost is nil, and a float32 mean of a product
+                # differs from the exact value by ~1e-5 relative — enough to
+                # move cond_por in the sixth decimal and to make an
+                # analytically-exact test fail for a reason that is not the
+                # model.
+                phi = float((cells.astype(np.float64) * block).mean())
             phi = self._clamp_por(phi)
             por.append(float(porosity_to_cond(phi, self.por_log_stats)))
             d, d6 = self._window_position(ov, box_lo, box_hi)
